7-extra-user-registration

+++ admin/includes/view_all_post.php
@@ -91,7 +91,7 @@ while($row = mysqli_fetch_assoc($edit_select_categories)) {
         echo "<td><a href='posts.php?draft=$post_id'>draft</a></td>";
         echo "<td><a href='../post.php?p_id=$post_id'>View Post</a></td>";
         echo "<td><a href='posts.php?source=edit_post&p_id=$post_id'>Edit</a></td>";
-        echo "<td><a href='posts.php?delete=$post_id'>Delete</a></td>";
+        echo "<td><a onClick=\"javascript: return confirm('Are you sure, you want to delete it?');\" href='posts.php?delete=$post_id'>Delete</a></td>";
         echo "</tr>";
     }
 ?>
